Changes:
- **orderBy** as a query param in the get publications endpoint
- Implementation of sorting and general refactoring
- Two new fields in PublicationBaseDataResponseDto
  - PublicationType
  - PublicationState
  - SubjectCategoryId

Example valid orderBy queries:
- AggregateRating, IncidentDate desc
- Title desc, Description
- AggregateRating desc

Example invalid orderBy queries:
- AggregateRating; IncidentDate desc
- Title Description desc

Supported properties for sorting:
- Title
- Description
- IncidentDate
- AggregateRating
- PublicationState
- PublicationType
- SubjectCategoryId

+++ Backend/PublicationService/src/LostAndFound.PublicationService.CoreLibrary/Internal/PaginationMetadata.cs
@@ -2,14 +2,14 @@
 {
     public class PaginationMetadata
     {
-        public int TotalItemCount { get; set; }
+        public long TotalItemCount { get; set; }
         public int TotalPageCount { get; set; }
         public int PageSize { get; set; }
         public int CurrentPage { get; set; }
         public string? NextPageLink { get; set; }
         public string? PreviousPageLink { get; set; }
 
-        public PaginationMetadata(int totalItemCount, int pageSize, int currentPage)
+        public PaginationMetadata(long totalItemCount, int pageSize, int currentPage)
         {
             TotalItemCount = totalItemCount;
             TotalPageCount = (int)Math.Ceiling(totalItemCount / (double)pageSize);
